requirements.txt
pandas>=1.3.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
ftfy>=6.0.0
regex>=2021.4.4
Pillow>=8.3.0
nltk>=3.6.0
scikit-learn>=0.24.0
torch-geometric>=2.0.0
